Make per-peer balance signed; negative pricing pays the peer
Negative pricing is a core economic mechanism (docs/design/core/
tollgate-pricing.md: "negative = node pays peer" to attract resources), but the
balance was u64 and cost_scaled clamped to zero — so a negative price silently
became "free." Make the accounting signed end to end.

- Price::cost_scaled returns i64 (no clamp): a negative rate yields a negative
  cost — we pay the peer rather than charge it.
- Session balance is now i64. Metering lowers it by the cost *magnitude*:
  charging (cost ≥ 0) draws down the peer's prepaid credit; paying (negative
  price) drives the balance below zero — the subsidy we owe. A peer we pay is
  never cut for exhaustion (suspension is gated on `charging`), so tolltop reads
  the sign as net earner (+) vs net spender (−), exactly the bidirectional case
  the docs describe ("I charge −1 sat/s for sending bytes" → net spender).
- PeerSnapshot.balance, status PeerStatus.their_balance/our_balance, the upstream
  consume loop's remaining, and ConsumeEvent.remaining_scaled all go i64; under
  negative pricing the consumer's remaining only grows, so it never tops up.

Positive pricing and suspension-on-exhaustion are unchanged (the magnitude equals
the cost when charging). New test: delivering at −2/unit drives the balance
negative and never suspends. Signed cost test replaces the old clamp test.

+++ crates/tollgate-net/src/client.rs
@@ -199,7 +199,9 @@ pub struct ConsumeEvent {
     pub peer_pubkey: String,
     pub price: Price,
     pub paid_scaled: u64,
-    pub remaining_scaled: u64,
+    /// Signed remaining balance: `+` we have prepaid credit left, `−` the
+    /// provider owes us (it delivers at a negative price — it pays us).
+    pub remaining_scaled: i64,
     /// The peer's latest report, if one arrived this poll.
     pub report: Option<ReportSummary>,
     /// The peer reported us balance-exhausted (cut off).
@@ -237,7 +239,7 @@ pub async fn run_consume(
         peer_pubkey: peer_pubkey.clone(),
         price,
         paid_scaled,
-        remaining_scaled: paid_scaled,
+        remaining_scaled: paid_scaled as i64,
         report: None,
         cut_off: false,
         topped_up: false,
@@ -288,8 +290,10 @@ pub async fn run_consume(
             .map_or(0, |r| price.cost_scaled(r.elapsed_ms, r.delivered));
 
         // Top up if cut off, or proactively before the balance dips below a top-up.
+        // Signed: under negative pricing `cost` is negative (the provider pays us),
+        // so remaining only grows and we never top up.
         let mut topped_up = false;
-        if cut_off || paid_scaled.saturating_sub(cost) < topup_scaled {
+        if cut_off || (paid_scaled as i64).saturating_sub(cost) < topup_scaled as i64 {
             let top = pay(base_url, identity, unit, mint_url, opts.topup_sat).await?;
             if top.accepted {
                 // A cut-off means the peer suspended us and reset its metering
@@ -309,7 +313,7 @@ pub async fn run_consume(
             peer_pubkey: peer_pubkey.clone(),
             price,
             paid_scaled,
-            remaining_scaled: paid_scaled.saturating_sub(cost),
+            remaining_scaled: (paid_scaled as i64).saturating_sub(cost),
             report: report.map(|r| ReportSummary {
                 elapsed_ms: r.elapsed_ms,
                 delivered: r.delivered,
